requirements.txt
Flask>=3.0.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0
opencv-python>=4.8.0
pydub>=0.25.1
matplotlib>=3.7.0
reportlab>=4.0.0
Werkzeug>=3.0.0

